Extend trimming to any list of indices
These indices are provided as a comma-separated list of numbers and
ranges, currently starting at zero. For example, the following are all
equivalent:

    1,2,3,4,5
    1-5
    1-3,4-5
    1-4,-1

In favor of more intuitive argument naming, the following changes have
been made to the command line interface:

- Add -e, --exclude argument to take indices, e.g. -1,2-3.
  The restrictions that previously applied to the --trim-last
  argument remain: it may not be used with --group-by at the same time.
- Remove -t, --trim-last argument; this can be achieved with -e -1.
- Remove --gs-profile short-form -g argument.
- Remove --im-profile short-form -i argument.
- Remove --im-quality short-form -q argument.
- Rename --group-by short-form -z argument to -g.
- Rename --batch, -x argument to --interactive, -i.

These changes hopefully make the interface more intuitive and easier
to remember usage. It is, as before, highly recommended to make
aliases with your preferred quality settings.